paso 9 mejorado

+++ 06-projects/items/gin/6/entity/entity.go
@@ -11,7 +11,7 @@ type Item struct {
 	Code        string
 	Title       string
 	Description string
-	Price       float64
+	Price       float32
 	Stock       int
 	Status      string
 	CreatedAt   time.Time
